chore: extract package validation to an extra script
and run it during CI
