+++ package.json
@@ -12,7 +12,7 @@
     "test:format": "pnpm run prettier --check",
     "test:lib": "pnpm --filter \"./packages/**\" run test:lib",
     "test:lib:dev": "pnpm --filter \"./packages/**\" run test:lib:dev",
-    "test:size": "pnpm run build && bundlewatch",
+    "test:build": "pnpm run build && bundlewatch && pnpm run validatePackages",
     "test:types": "pnpm --filter \"./packages/**\" run test:types",
     "build": "rollup --config rollup.config.js && pnpm --filter \"./packages/**\" run build && pnpm run build:types",
     "build:types": "pnpm --filter \"./packages/**\" run build:types",
@@ -20,7 +20,8 @@
     "dev": "pnpm run watch",
     "prettier": "prettier --plugin-search-dir . \"{packages,examples}/**/src/**/*.{md,js,jsx,ts,tsx,json,vue,svelte}\"",
     "prettier:write": "pnpm run prettier --write",
-    "cipublish": "ts-node scripts/publish.ts"
+    "cipublish": "ts-node scripts/publish.ts",
+    "validatePackages": "ts-node scripts/validate-packages.ts"
   },
   "namespace": "@tanstack",
   "devDependencies": {
